imgAnalysis/requirements.txt
# requirements.txt

flask_restful
requests
flask
google.cloud.vision